Move dagger files into separate package within dependency injection.
